requirements.txt
streamlit
python-dotenv
google-genai
google-generativeai
crewai
crewai-tools
reportlab